backend/meeting_service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-JWT-Extended
pydantic
redis
flask-cors 